Improve Lahza card SMS details
