+++ server/routes/orders.js
@@ -21,7 +21,7 @@ const {
   hasAnyTranslation,
   mapLocalizedForResponse,
 } = require("../utils/localized");
-const { sendSMSHTD, normalizePhone } = require("../utils/sms");
+const { toDisplayName, sendOrderConfirmationSMS } = require("../utils/order-sms");
 const DEFAULT_RECAPTCHA_ACTION = "checkout";
 const ENV_MIN_SCORE = Number(process.env.RECAPTCHA_MIN_SCORE);
 const DEFAULT_RECAPTCHA_MIN_SCORE = Number.isFinite(ENV_MIN_SCORE)
@@ -124,11 +124,6 @@ const toOid = (id) =>
   mongoose.isValidObjectId(id) ? new mongoose.Types.ObjectId(id) : null;
 const slugify = (s) =>
   (s || "").toString().trim().toLowerCase().replace(/\s+/g, "-");
-
-const toDisplayName = (raw) => {
-  const normalized = ensureLocalizedObject(raw);
-  return normalized.ar || normalized.he || "منتج";
-};
 
 const resolveItemName = ({
   requestedName,
@@ -297,201 +292,6 @@ async function computeOrderTotals(items, incomingDiscount) {
   const total = Math.max(0, subtotal - discount.amount);
 
   return { subtotal, discount, total };
-}
-
-const MAX_SMS_ITEMS = 4;
-const ADDRESS_SMS_MAX_LENGTH = 70;
-
-const CARD_TYPE_LABELS = {
-  visa: "فيزا",
-  mastercard: "ماستركارد",
-  master: "ماستركارد",
-  maestro: "مايسترو",
-  amex: "أمريكان إكسبريس",
-  americanexpress: "أمريكان إكسبريس",
-  diners: "داينرز كلوب",
-  dinersclub: "داينرز كلوب",
-  discover: "ديسكفر",
-  jcb: "جي سي بي",
-  mada: "مدى",
-  unionpay: "يونيون باي",
-};
-
-function describePaymentMethod(method) {
-  if (!method) return "";
-  const normalized = String(method).trim().toLowerCase();
-  if (!normalized) return "";
-  if (normalized === "card") return "الدفع بالبطاقة";
-  if (normalized === "cod") return "الدفع عند التوصيل";
-  return String(method);
-}
-
-function translateCardTypeForSms(cardType) {
-  if (cardType == null) return "";
-  const raw = String(cardType).trim();
-  if (!raw) return "";
-  if (/[ء-ي]/.test(raw)) return raw;
-  const key = raw.toLowerCase().replace(/[^a-z]/g, "");
-  return CARD_TYPE_LABELS[key] || raw;
-}
-
-function formatCardLast4ForSms(last4) {
-  if (last4 == null) return "";
-  const digits = String(last4).replace(/\D/g, "").slice(-4);
-  if (!digits) return "";
-  return `****${digits}`;
-}
-
-function formatOrderDateForSms(createdAt) {
-  if (!createdAt) return "";
-  const date = createdAt instanceof Date ? createdAt : new Date(createdAt);
-  if (!(date instanceof Date) || Number.isNaN(date.getTime())) return "";
-  const pad = (n) => String(n).padStart(2, "0");
-  const y = date.getFullYear();
-  const m = pad(date.getMonth() + 1);
-  const d = pad(date.getDate());
-  const hrs = pad(date.getHours());
-  const mins = pad(date.getMinutes());
-  return `${y}-${m}-${d} ${hrs}:${mins}`;
-}
-
-function truncateText(text, maxLength) {
-  if (!text) return "";
-  const value = String(text);
-  if (value.length <= maxLength) return value;
-  const sliceEnd = Math.max(0, maxLength - 1);
-  return `${value.slice(0, sliceEnd).trimEnd()}…`;
-}
-
-function formatOrderTotal(amount) {
-  const numeric = Number(amount);
-  if (!Number.isFinite(numeric)) return "0";
-  const rounded = Math.round(numeric * 100) / 100;
-  if (Number.isInteger(rounded)) return String(rounded);
-  return rounded.toFixed(2);
-}
-
-function formatOrderItemsForSms(items) {
-  if (!Array.isArray(items)) return [];
-  return items.slice(0, MAX_SMS_ITEMS).map((item) => {
-    const name = truncateText(toDisplayName(item?.name), 32);
-    const qty = Math.max(1, parseInt(item?.quantity, 10) || 1);
-    const suffix = qty > 1 ? ` x${qty}` : "";
-    return `• ${name}${suffix}`;
-  });
-}
-
-function buildOrderSmsMessage({
-  orderId,
-  items,
-  total,
-  currency,
-  address,
-  paymentMethod,
-  paymentCardType,
-  paymentCardLast4,
-  orderCreatedAt,
-}) {
-  if (!orderId) return "";
-  const lines = [`شكراً لطلبك من ديكوري!`, `رقم الطلب: ${orderId}`];
-
-  const itemLines = formatOrderItemsForSms(items);
-  if (itemLines.length) {
-    lines.push("المنتجات:");
-    lines.push(...itemLines);
-    if (Array.isArray(items) && items.length > MAX_SMS_ITEMS) {
-      const remaining = items.length - MAX_SMS_ITEMS;
-      if (remaining === 1) {
-        lines.push("و1 منتج إضافي...");
-      } else if (remaining > 1) {
-        lines.push(`و${remaining} منتجات إضافية...`);
-      }
-    }
-  }
-
-  const safeCurrency = currency || DEFAULT_PAY_CURRENCY;
-  lines.push(`الإجمالي: ${formatOrderTotal(total)} ${safeCurrency}`);
-
-  const trimmedAddress = truncateText(String(address || "").trim(), ADDRESS_SMS_MAX_LENGTH);
-  if (trimmedAddress) {
-    lines.push(`العنوان: ${trimmedAddress}`);
-  }
-
-  const methodDescription = describePaymentMethod(paymentMethod);
-  const normalizedMethod = String(paymentMethod || "")
-    .trim()
-    .toLowerCase();
-  if (methodDescription) {
-    lines.push(`طريقة الدفع: ${methodDescription}`);
-  }
-
-  if (normalizedMethod === "card") {
-    const cardTypeText = translateCardTypeForSms(paymentCardType);
-    lines.push(`نوع البطاقة: ${cardTypeText || "غير متوفرة"}`);
-    const last4Text = formatCardLast4ForSms(paymentCardLast4);
-    lines.push(`آخر 4 أرقام من البطاقة: ${last4Text || "غير متوفرة"}`);
-  }
-
-  const orderDateText = formatOrderDateForSms(orderCreatedAt);
-  if (orderDateText) {
-    lines.push(`تاريخ الطلب: ${orderDateText}`);
-  }
-
-  lines.push("سنتواصل معك لتأكيد الطلب.");
-  return lines.join("\n");
-}
-
-function pickOrderPhone(userObj, guestObj) {
-  if (userObj?.phone) return userObj.phone;
-  if (guestObj?.phone) return guestObj.phone;
-  return null;
-}
-
-function isLikelySmsPhone(phone) {
-  if (typeof phone !== "string") return false;
-  const digits = phone.replace(/\D/g, "");
-  return digits.length >= 9;
-}
-
-async function sendOrderConfirmationSMS({ order, items, user, guest }) {
-  if (!order || !order._id) return;
-
-  const candidatePhone = pickOrderPhone(user, guest);
-  const normalized = normalizePhone(candidatePhone);
-  if (!isLikelySmsPhone(normalized)) return;
-
-  const message = buildOrderSmsMessage({
-    orderId: String(order._id),
-    items,
-    total: Number(order.total || 0),
-    currency: order.paymentCurrency || DEFAULT_PAY_CURRENCY,
-    address: order.address,
-    paymentMethod: order.paymentMethod,
-    paymentCardType: order.paymentCardType,
-    paymentCardLast4: order.paymentCardLast4,
-    orderCreatedAt: order.createdAt,
-  });
-
-  if (!message) return;
-
-  try {
-    const smsResult = await sendSMSHTD(normalized, message, {
-      label: `order:${order._id}`,
-    });
-    if (!smsResult?.ok) {
-      console.error("Order confirmation SMS did not send successfully", {
-        orderId: order._id,
-        phone: normalized,
-        result: smsResult,
-      });
-    }
-  } catch (smsErr) {
-    console.error("Failed to send order confirmation SMS", {
-      orderId: order._id,
-      phone: normalized,
-      error: smsErr?.response?.data || smsErr?.message || smsErr,
-    });
-  }
 }
 
 /* ======================= إنشاء طلب COD ======================= */
@@ -968,6 +768,18 @@ router.patch(
           )
         )
       );
+
+      await sendOrderConfirmationSMS({
+        order: updated,
+        items: updated.items,
+        user: updated.user,
+        guest: updated.isGuest ? updated.guestInfo : undefined,
+        overrides: {
+          paymentCardType: updateSet.paymentCardType || verification.cardType,
+          paymentCardLast4:
+            updateSet.paymentCardLast4 || verification.cardLast4,
+        },
+      });
 
       return res.json(updated);
     } catch (err) {
